requirements.txt
pandas
numpy
python-dotenv
nbformat
requests
web3